Module Comments_state.Comment_node

type 'a t = 'a comment_node

include Ppx_compare_lib.Equal.S1 with type 'a t := 'a t

val equal : 'a Base__Ppx_compare_lib.equal -> 'a t Base__Ppx_compare_lib.equal

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

val return : 'a -> 'a t

val value_exn : 'a t -> 'a

Once attach_comments has been called, all comments are attached to nodes, and all functions f supplied to comment_node are executed. If attach_comments hasn't been called, this function will always raise.